Tests/LibWeb/Text/expected/recompute-inherited-style-of-display-none-descendant.txt

Visibility before: hidden
Color      before: rgb(255, 0, 0)
Visibility after: hidden
Color      after: rgb(255, 0, 0)
